upgrade to C++23 with CMake, fix warnings, add KNN and logistic regression
- Add CMakeLists.txt targeting C++23
- Fix uninitialized epsilon in KernelSVM
- Fix memory leak in DecisionTree using unique_ptr
- Fix duplicate euclideanDistance between k_means and tsne
- Fix typo in matthews_correlation_coefficient
- Replace typedef with using aliases
- Replace cout/cerr with std::println
- Replace M_PI with std::numbers::pi
- Add structured bindings, virtual destructors
- Fix all -Wall -Wextra -Wpedantic -Wsign-conversion -Wconversion warnings
- Add KNNClassifier and KNNRegressor
- Add LogisticRegression
- Add .gitignore entries for build artifacts
- Run clang-format on all source files

+++ supervised/mlp.cpp
@@ -28,10 +28,7 @@ Matrix applyFunction(const Matrix &mat, double (*func)(double)) {
   return result;
 }
 
-double sigmoid_derivative(double x) {
-  return x * (1.0 -
-              x);
-}
+double sigmoid_derivative(double x) { return x * (1.0 - x); }
 
 class MLP {
 private:
@@ -42,7 +39,7 @@ private:
   double learningRate = 0.01;
 
 public:
-  MLP(int inputSize, int hiddenSize, int outputSize) {
+  MLP(size_t inputSize, size_t hiddenSize, size_t outputSize) {
     weightsInputToHidden = Matrix(inputSize, Vector(hiddenSize));
     weightsHiddenToOutput = Matrix(hiddenSize, Vector(outputSize));
     hiddenBias = Vector(hiddenSize, 0.1);
@@ -58,6 +55,7 @@ public:
 
     return outputActivated;
   }
+
   void train(const Vector &input, const Vector &targetOutput) {
     Matrix hidden = multiply({input}, weightsInputToHidden);
     Vector hiddenActivated = applyFunction(hidden[0], sigmoid);
